requirements.txt
Flask
flask_restx
SQLAlchemy
sqlalchemy-utils
pymysql
cryptography
mlflow
flask_cors
psutil